Update
I have added code that allows you to take snippets of video with the camera. I was going to add something to skip the GPS if it took to long to find data, but that is going to require threading or multiprocessing or some crap like that. Not today, my friend.